+++ grid2op/Observation/BaseObservation.py
@@ -1150,7 +1150,7 @@ class BaseObservation(GridObjects):
         lex_bus, lex_conn = self._get_bus_id(self.line_ex_pos_topo_vect, self.line_ex_to_subid)
 
         if self.shunts_data_available:
-            sh_bus = self._shunt_bus
+            sh_bus = 1 * self._shunt_bus
             sh_bus[sh_bus > 0] = self.shunt_to_subid[sh_bus > 0]*(sh_bus[sh_bus > 0] - 1) + \
                                  self.shunt_to_subid[sh_bus > 0]
             sh_conn = self._shunt_bus != -1